CODE-164 Remove old scripts which are no longer used
