--- sous_chef/helper_scripts/cleanup_recipe/standardize_unit_name.py
@@ -1,2 +1,0 @@
-def standardize_unit_names():
-    pass
