--- sous_chef/menu_template/base_email.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-        "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" lang="en-GB">
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-    <title>Food plan</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <style>
-        #checkboxes label {
-            float: left;
-        }
-
-        #checkboxes ul {
-            margin: 0;
-            list-style: none;
-            float: left;
-        }
-    </style>
-</head>
-<body>
-<div id="checkboxes">
-    <ul class="no-bullets">
-        <li> Review (& correct) automated plan</li>
-        <ul class="no-bullets">
-            <li><input type="checkbox"> Are there calendar conflicts?</li>
-            <li><input type="checkbox"> Correct veggie side dishes?</li>
-            <li><input type="checkbox"> Add starch side dish?</li>
-            <li><input type="checkbox"> Should we pad with a defrosted meal?</li>
-        </ul>
-        <li><input type="checkbox"> Create shopping list</li>
-        <li><input type="checkbox"> Transcribe plan to RecetteTek/food calendar</li>
-        <li><input type="checkbox"> Add defrost tasks</li>
-    </ul>
-</div>
-<br><br><br><br><br><br><br><br><br><br>
-</body>
-</html>
