--- sous_chef/menu_template/cuisine_map.yml
@@ -1,38 +1,0 @@
-Mediterranean:
-  - Greek
-  - Italian
-  - Mediterranean
-
-Asian:
-  - Asian
-  - Chinese
-  - Japanese
-  - Korean
-  - Thai
-  - Vietnamese
-
-Indian:
-  - Indian
-
-Latin:
-  - Argentinian
-  - Brazilian
-  - Latin food
-  - Mexican
-  - Portuguese
-  - Spanish
-  - Tex Mex
-
-Middle East:
-  - Israeli
-  - Middle East
-  - Moroccan
-  - Syrian
-
-Other:
-  - American
-  - British
-  - comfort food
-  - French
-  - German
-  - Irish
