+++ sous_chef/tests/unit_tests/recipe_book/test_read_recipe_book.py
@@ -9,6 +9,7 @@ from sous_chef.recipe_book.read_recipe_book import (
     Recipe,
     RecipeBook,
     SelectRandomRecipeError,
+    create_timedelta,
 )
 from tests.util import assert_equal_dataframe
 
@@ -225,3 +226,23 @@ class TestRecipeBook:
             assert_equal_dataframe(recipe_book.dataframe, recipe1)
         else:
             assert_equal_dataframe(recipe_book.dataframe, recipe2)
+
+
+@pytest.mark.parametrize(
+    "input_time_string,expected_timedelta",
+    [
+        ("hurr 30 min", pd.to_timedelta("00:30:00")),
+        ("20 min 10 s", pd.to_timedelta("00:20:10")),
+        ("0:10", pd.to_timedelta("00:10:00")),
+        ("0:10:0", pd.to_timedelta("00:10:00")),
+        ("0:10:00", pd.to_timedelta("00:10:00")),
+        ("00:10:00", pd.to_timedelta("00:10:00")),
+        ("00:10:0", pd.to_timedelta("00:10:00")),
+        ("5 hours", pd.to_timedelta("05:00:00")),
+        ("15 minutes", pd.to_timedelta("00:15:00")),
+        ("5 hours 10 mins", pd.to_timedelta("05:10:00")),
+        ("prep time 6 min", pd.to_timedelta("00:06:00")),
+    ],
+)
+def test_create_timedelta(input_time_string, expected_timedelta):
+    assert create_timedelta(input_time_string) == expected_timedelta
